+++ src/main/java/ru/spliterash/vkVideoUnlocker/group/api/Groups.kt
@@ -4,8 +4,8 @@ import ru.spliterash.vkVideoUnlocker.group.dto.GroupInfo
 import ru.spliterash.vkVideoUnlocker.group.vkModels.LongPollServerResponse
 
 interface Groups {
-    suspend fun status(groupId: Int): GroupInfo
-    suspend fun join(groupId: Int)
+    suspend fun status(groupId: Long): GroupInfo
+    suspend fun join(groupId: Long)
 
     suspend fun getLongPollServer(): LongPollServerResponse
