+++ src/main/java/ru/spliterash/vkVideoUnlocker/video/holder/AbstractVideoContentHolder.kt
@@ -3,16 +3,16 @@ package ru.spliterash.vkVideoUnlocker.video.holder
 import ru.spliterash.vkVideoUnlocker.video.dto.FullVideo
 
 abstract class AbstractVideoContentHolder(
-    val contentId:String
+    val contentId: String
 ) : VideoContentHolder {
     protected var fullVideo: FullVideo? = null
 
-    override val ownerId: Int
+    override val ownerId: Long
         get() {
             val contentId = contentId
             val split = contentId.split("_")
 
-            return split[0].toInt()
+            return split[0].toLong()
         }
 
     protected abstract suspend fun loadFullVideo(): FullVideo
